sdk version update, shortform readme added
